Extract InlineCDataCreatorTrait for shared createInlineDereferencable
Unify duplicated createInlineZval/createInlineZendArray methods across
8 classes into a single generic createInlineDereferencable(field, class)
method in a shared trait. The trait also provides setPointedTypeResolver
and the $pointed_type_resolver property, eliminating identical
boilerplate from all implementing classes.

Changes:
- New InlineCDataCreatorTrait with createInlineDereferencable() and
  setPointedTypeResolver()
- Updated 8 classes to use the trait: ZendExecutorGlobals,
  ZendClassEntry, ZendCompilerGlobals, ZendExecuteData,
  ZendClassConstant, ZendReference, ZendClosure, ZendConstant,
  ZvalArray
- Normalized ZendConstant offset calculation and ZendClosure size
  calculation to use the standard pattern
