+++ src/Lib/PhpInternals/Types/Zend/ZendClassConstant.php
@@ -31,7 +31,7 @@ class ZendClassConstant implements Dereferencable, PointedTypeResolverAware
     /** @var Pointer<ZendString>|null */
     public ?Pointer $doc_comment;
 
-    private ?PointedTypeResolver $pointed_type_resolver = null;
+    use InlineCDataCreatorTrait;
 
     /** @var Pointer<ZendArray>|null */
     public ?Pointer $attributes;
@@ -60,7 +60,7 @@ class ZendClassConstant implements Dereferencable, PointedTypeResolverAware
     public function __get(string $field_name): mixed
     {
         return match ($field_name) {
-            'value' => $this->value = $this->createInlineZval(),
+            'value' => $this->value = $this->createInlineDereferencable('value', Zval::class),
             'doc_comment' => $this->doc_comment = $this->casted_cdata->casted->doc_comment !== null
                 ? Pointer::fromCData(
                     ZendString::class,
@@ -88,32 +88,6 @@ class ZendClassConstant implements Dereferencable, PointedTypeResolverAware
         };
     }
 
-
-    #[\Override]
-    public function setPointedTypeResolver(PointedTypeResolver $resolver): void
-    {
-        $this->pointed_type_resolver = $resolver;
-    }
-
-    private function createInlineZval(): Zval
-    {
-        $zval_class = $this->pointed_type_resolver !== null
-            ? $this->pointed_type_resolver->resolve(Zval::class)
-            : Zval::class;
-        return $zval_class::fromCastedCData(
-            new CastedCData(
-                $this->casted_cdata->casted->value,
-                $this->casted_cdata->casted->value,
-            ),
-            new Pointer(
-                $zval_class,
-                $this->pointer->address
-                +
-                \FFI::typeof($this->casted_cdata->casted)->getStructFieldOffset('value'),
-                \FFI::sizeof($this->casted_cdata->casted->value),
-            ),
-        );
-    }
 
     #[\Override]
     public static function getCTypeName(): string
