+++ src/Lib/PhpInternals/Types/Zend/ZendReference.php
@@ -31,13 +31,7 @@ class ZendReference implements Dereferencable, PointedTypeResolverAware
     /** @psalm-suppress PropertyNotSetInConstructor */
     public Zval $val;
 
-    private ?PointedTypeResolver $pointed_type_resolver = null;
-
-    #[\Override]
-    public function setPointedTypeResolver(PointedTypeResolver $resolver): void
-    {
-        $this->pointed_type_resolver = $resolver;
-    }
+    use InlineCDataCreatorTrait;
 
     /**
      * @param CastedCData<\FFI\PhpInternals\zend_reference> $casted_cdata
@@ -55,28 +49,8 @@ class ZendReference implements Dereferencable, PointedTypeResolverAware
     {
         return match ($field_name) {
             'gc' => $this->gc = new ZendRefcountedH($this->casted_cdata->casted->gc),
-            'val' => $this->val = $this->createInlineZval(),
+            'val' => $this->val = $this->createInlineDereferencable('val', Zval::class),
         };
-    }
-
-    private function createInlineZval(): Zval
-    {
-        $zval_class = $this->pointed_type_resolver !== null
-            ? $this->pointed_type_resolver->resolve(Zval::class)
-            : Zval::class;
-        return $zval_class::fromCastedCData(
-            new CastedCData(
-                $this->casted_cdata->casted->val,
-                $this->casted_cdata->casted->val,
-            ),
-            new Pointer(
-                $zval_class,
-                $this->pointer->address
-                +
-                \FFI::typeof($this->casted_cdata->casted)->getStructFieldOffset('val'),
-                \FFI::sizeof($this->casted_cdata->casted->val),
-            ),
-        );
     }
 
     #[\Override]
